refactoring fake docker handler
